Added current booking feature to HomeModel
A new property `CurrentBooking` of type `BookingDto` has been added to the `HomeModel` class in `Home.cshtml.cs` under the `RSAllies.Pages` namespace. This property will hold the current booking information for the user. The `OnGetAsync` method of the `HomeModel` class now includes a call to `apiClient.GetCurrentUserBooking(UserData!.Id)` to fetch the current booking of the user. If the result is not null, it is assigned to the `CurrentBooking` property. No actual changes were made to `Index.cshtml.cs` under the `RSAllies.Contracts` namespace, only formatting or indentation changes.