Added new properties, methods, and services, and updated existing ones
In this commit, we have made several changes across multiple files. We added `UserId` and `Name` properties to the `CreateCertificate` class and `Name` property to the `UserResponseDto` record. The `Mark` method in `MarkingService.cs` was updated to accept nullable `List<AnswerDto>`. In `Program.cs`, we added several using statements, modified the `UseWolverine` method, added a `PublishMessage` method, and added several services to the DI container. We also added a `DownloadFile` method in `FilesController.cs` and a `Handle` method in `CreateCertificateHandler.cs`. In `UserResponseDtoHandler.cs`, we added several using statements, changed the namespace, added a constructor, a `GetAnswers` method, and a `Handle` method. We added references and image files to `RSAllies.PDFService.csproj`. Lastly, we updated `English.cshtml.cs` and `Swahili.cshtml.cs` to accept a `UserResponsePublisher` in the constructor, added a `Name` property to the `UserResponseDto`, and uncommented a line of code to publish user responses.