Update file path and add error handling in Search.cshtml.cs
The commit includes two major changes. Firstly, the path of the file from which JSON is read in `MarkingService.cs` has been updated. It now reads from the `RSAllies.MarkingService` directory instead of the `RSAllies.PDFService` directory. Secondly, in `Search.cshtml.cs`, the assignment of the `Sessions` variable is now wrapped in a conditional statement. This checks if the result of `GetFilteredSessionsAsync` is successful before assigning `Sessions` the value of `result.Value`, adding a layer of error handling to the code.