Updated `qodana.yaml` with new configurations
Updated the `qodana.yaml` file to enhance code analysis. Added a comment block for file description and a link to JetBrains help page. Set `version` to "1.0" and `ide` to "QDNET" for CI/CD pipeline compatibility. Specified "qodana.starter" as the inspection profile in the `profile` field. Added an `exclude` field to omit certain paths from the analysis.